🚑 fix: auth dio revert

+++ lib/apis/auth_dio.dart
@@ -20,20 +20,14 @@ Future<Dio> authDio(BuildContext context) async {
     return handler.next(options);
   }, onError: (error, handler) async {
     debugPrint("authDio Error: ${error.response?.data.toString()}");
-    await storage.deleteAll();
-    Navigator.pushAndRemoveUntil(
-      context,
-      MaterialPageRoute(builder: (context) => const LoginScreen()),
-      (route) => false,
-    );
-    // if (error.response?.statusCode == 401) {
-    //   await storage.deleteAll();
-    //   Navigator.pushAndRemoveUntil(
-    //     context,
-    //     MaterialPageRoute(builder: (context) => const LoginScreen()),
-    //     (route) => false,
-    //   );
-    // }
+    if (error.response?.statusCode == 401) {
+      await storage.deleteAll();
+      Navigator.pushAndRemoveUntil(
+        context,
+        MaterialPageRoute(builder: (context) => const LoginScreen()),
+        (route) => false,
+      );
+    }
     return handler.next(error);
   }));
   return dio;
